Move remaining shared code to the core StatWeights.js
